Azure/CT/requirements.txt
azure-core==1.26.3
azure-storage-blob==12.15.0
scikit-learn==1.2.2
pandas==1.4.4
numpy==1.22.4
Pillow==8.4.0
